chore: 整理目录, 源码使用ts开发
