--- packages/sdk-ts/CHANGELOG.md
@@ -1,52 +1,0 @@
-# sharding-download
-
-## 1.2.2
-
-### Patch Changes
-
-- chore: 打包声明文件
-
-## 1.2.1
-
-### Patch Changes
-
-- fix: 默认db name和version
-
-## 1.2.0
-
-### Minor Changes
-
-- typescript 重构
-
-## 1.1.0
-
-### Minor Changes (2025-05-28)
-
-#### 新增
-
-- storageVersion: db name
-- storageName: db version
-
-## 1.0.2
-
-### Patch Changes
-
-- fix: resume操作时需要取消getMeta接口请求
-
-## 1.0.1
-
-### Patch Changes
-
-- chore: keywords
-
-## 1.0.0
-
-### Major Changes
-
-- readme更新
-
-## 0.0.2
-
-### Patch Changes
-
-- 大文件分片下载
