+++ packages/sdk/examples/quick-start.js
@@ -1,5 +1,6 @@
-import Downloader, { FileStatus, Callbacks } from '../src/index.js'
-import { customConfig, urlOptions, actionList } from './options.js'
+import { Downloader, FileStatus, Callbacks } from '../src/index'
+// import { Downloader, FileStatus, Callbacks } from '../dist/sharding-downloader.es.js'
+import { customConfig, urlOptions, actionList } from './options'
 
 const { ref, reactive, watch } = Vue
