+++ packages/sdk/src/core/constants.ts
