+++ packages/sdk/src/core/typings/downloader.type.ts
