--- packages/sdk/src/index.js
@@ -1,6 +1,0 @@
-import Downloader from './Downloader'
-import * as Utils from './utils'
-import { FileStatus, Callbacks } from './constants'
-
-export default Downloader
-export { Utils, FileStatus, Callbacks }
