--- packages/sdk/vite.config.js
@@ -1,22 +1,0 @@
-import path from 'node:path'
-import { defineConfig } from 'vite'
-
-export default defineConfig({
-  resolve: {
-    alias: {
-      '@': path.join(__dirname, 'src')
-    }
-  },
-  server: {
-    port: 3200,
-    open: '/examples/quick-start.html'
-  },
-  build: {
-    lib: {
-      entry: './src/index.js',
-      name: 'ShardingDownload',
-      fileName: 'sharding-download'
-    },
-    minify: true
-  }
-})
